MorphoFeatures/requirements_lightning.txt
torch>=1.10.0
pytorch-lightning>=1.6.0
monai>=0.9.0
pyyaml>=6.0
numpy>=1.20.0
h5py>=3.7.0
z5py>=2.0.16
scikit-image>=0.19.0
pandas>=1.3.0
pybdv>=0.5.0 